% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{adjacent_polys}
\alias{adjacent_polys}
\title{Calculate adjacency list for a shape file.}
\usage{
adjacent_polys(shp, add_adj_for_islands = T, nearest_for_islands = 1)
}
\arguments{
\item{shp}{A sf object}

\item{add_adj_for_islands}{Boolean. Default is \code{TRUE}. If there are unconnected islands, add adjancecies to nearest polygon(s).}

\item{nearest_for_islands}{Integer. Default is \code{1}. How many adjacencies to add for islands.}
}
\value{
A dataframe of the indices of adjacent polygons.
}
\description{
Calculate adjacency list for a shape file.
}
\examples{

}
