% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bos}
\alias{bos}
\title{A color palette based on the City of Boston Brand Guidelines
https://www.boston.gov/departments/innovation-and-technology/brand-guidelines}
\format{
An object of class \code{list} of length 10.
}
\usage{
bos
}
\description{
A color palette based on the City of Boston Brand Guidelines
https://www.boston.gov/departments/innovation-and-technology/brand-guidelines
}
\keyword{datasets}
