% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{centroid_of_inscribed_circle}
\alias{centroid_of_inscribed_circle}
\title{Find centroid of the inscribed circle of a polygon}
\usage{
centroid_of_inscribed_circle(shp)
}
\arguments{
\item{shp}{A polygon shape file}
}
\value{
List of points
}
\description{
Find centroid of the inscribed circle of a polygon
}
