% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{color_shp}
\alias{color_shp}
\title{Color polygons of a map}
\usage{
color_shp(shp, n_cols = 5, retry = TRUE, ...)
}
\arguments{
\item{shp}{A polygon shape file}

\item{n_cols}{The number of colors to use. Defaul is 5.}

\item{retry}{If the algorithm fails, retry with \code{n_cols+1} colors. Default is TRUE. Function fails if polygons unable to be colors with 10 colors.}

\item{...}{Additional arguments, only used internally.}
}
\value{
A vector of numbers identifying the colors for each shape.
}
\description{
This function assigns colors to polygons in a map to (1) color each polygon differently from its neighbors, and (2) use each color roughly the same number of times.
}
