% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{coord_sf_from_sf}
\alias{coord_sf_from_sf}
\title{Set ggplot coordinates using bounding box of sf object}
\usage{
coord_sf_from_sf(shp, expand = TRUE, ...)
}
\arguments{
\item{shp}{A sf object.}

\item{expand}{A boolean if the coordinates should be expanded to include padding around the shape.}

\item{...}{Additional arguments passed to \code{coord_sf}.}
}
\description{
Set ggplot coordinates using bounding box of sf object
}
