% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_name.R
\name{date_name}
\alias{date_name}
\title{date_name}
\usage{
date_name(
  date,
  formal = F,
  year = F,
  month.full = F,
  weekday = F,
  weekday.parens = F
)
}
\arguments{
\item{date}{A date object}

\item{formal}{Boolean. Default is FALSE.}

\item{year}{Boolean. Default is FALSE.}

\item{month.full}{Boolean. Default is FALSE.}

\item{weekday}{Boolean. Default is FALSE.}

\item{weekday.parens}{Boolean. Default is FALSE.}
}
\value{
A string of the date name
}
\description{
date_name
}
