% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_na_cols}
\alias{drop_na_cols}
\title{Drop columns of a data frame where all values are NA}
\usage{
drop_na_cols(x, verbose = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{verbose}{Optional. Display list of dropped columns.}
}
\value{
A data frame, without any columns where all values are missing.
}
\description{
Drop columns of a data frame where all values are NA
}
\examples{
x <- data.frame(a=1:5, b=1:5, c=NA, d=NA, e=letters[1:5])
drop_na_cols(x, verbose=TRUE)
}
