% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei_output.R
\name{ei_output}
\alias{ei_output}
\title{Ecological Inference Output Processing}
\usage{
ei_output(
  ei_md_model,
  cols = NULL,
  formula = NULL,
  full = FALSE,
  drop_last_col = FALSE
)
}
\arguments{
\item{ei_md_model}{A model object estimated from eiPack::ei.MD.bayes()}

\item{cols}{A vector of column names to include in the output, as characters. If NULL, columns are automatically detected from the model formula.}

\item{formula}{The formula usef for running EI. Must be provided if cols is null.}

\item{full}{If TRUE, provide differences between columns across rows and between rows across columns. Default is FALSE.}

\item{drop_last_col}{If TRUE, drop the last column variable. Only applies if cols is not supplied. Default is FALSE.}
}
\value{
If full==FALSE, a data frame with means and confidence intervals for each row and specified column. If full==TRUE, a named
list with (1) a data frame with means and confidence intervals for each row and specified column, (2) a data frame with means and CIs for
differences between each pair of rows for every column, (3) a data frame with means and CIs for the differences between each pair of columns
for every row, (4) the number of ei.samples, and (5) the number of observations in the data.
}
\description{
Ecological Inference Output Processing
}
