% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fips}
\alias{fips}
\title{List of County FIPS Codes}
\format{
An object of class \code{data.frame} with 3245 rows and 5 columns.
}
\source{
Kyle Walker / Tigris Package
}
\usage{
fips
}
\description{
List of County FIPS Codes
}
\keyword{datasets}
