% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{geom_sf_label2}
\alias{geom_sf_label2}
\alias{geom_sf_text2}
\title{Add geom_sf_label with labels located in inscribed circles instead of points on surface.}
\usage{
geom_sf_label2(...)

geom_sf_text2(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{geom_sf_label}.}
}
\description{
Add geom_sf_label with labels located in inscribed circles instead of points on surface.
}
