% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letter_grades.R
\name{letter_grades}
\alias{letter_grades}
\title{Calculate letter grades from percentages}
\usage{
letter_grades(
  scores,
  vals = c(0.93, 0.9, 0.87, 0.83, 0.8, 0.77, 0.73, 0.7, 0.6, 0),
  labs = c("A", "A-", "B+", "B", "B-", "C+", "C", "C-", "D", "F")
)
}
\arguments{
\item{scores}{A vector of percentages}

\item{vals}{Optional. A vector of the lower bound of each grade level, sorted from best to worst.}

\item{labs}{Optional. A vector of the names of each grade, sorted from best to worse. Must be the same length as vals.}
}
\value{
A vector of grades, with factor levels as specified by labs.
}
\description{
Calculate letter grades from percentages
}
\examples{
letter_grades(c(.6, .75, .85, .95, 1))
letter_grades(scores=c(.5, .75, .9, .3),
              vals=c(.8, .6, .4, 0),
              labs=c("Excellent", "Good", "Poor", "Fail"))

}
