% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_of_lists_to_tibble}
\alias{list_of_lists_to_tibble}
\title{List of Lists to Tibble}
\usage{
list_of_lists_to_tibble(x)
}
\arguments{
\item{x}{A list of lists. Each sublist must contain items with the same names and types.}
}
\value{
A tibble
}
\description{
List of Lists to Tibble
}
