% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ma}
\alias{ma}
\title{Shape file of Massachusetts counties}
\format{
A sf object with 14 rows and 10 variables.
}
\source{
U.S. Census
}
\usage{
ma
}
\description{
Shape file of Massachusetts counties
}
\keyword{datasets}
