% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mass_munis}
\alias{mass_munis}
\title{Massachusetts Municipalities}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 16499 rows and 6 columns.
}
\source{
US Census shape file via tigris
}
\usage{
mass_munis
}
\description{
Massachusetts Municipalities
}
\keyword{datasets}
