% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mcols}
\alias{mcols}
\title{Custom Colors}
\format{
An object of class \code{list} of length 220.
}
\source{
Tailwind CSS https://tailwindcss.com/docs/customizing-colors
}
\usage{
mcols
}
\description{
Custom Colors
}
\keyword{datasets}
