% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmer.R
\name{palmer}
\alias{palmer}
\title{PALMER}
\usage{
palmer(X, path = NULL, K = 3, L = 3, B = 1000)
}
\arguments{
\item{X}{Matrix of binary values with n samples by p variables}

\item{path}{List. Gene list for the prior information (default: NULL)}

\item{K}{The number of gene clusters (default: 3)}

\item{L}{The number of GO term clusters (default: 3)}

\item{B}{The number of bootstrappings (default: 1000)}
}
\description{
A constrained biclustering Algoritm to improve pathway annotation bssed on the biomedical literature mining
}
\examples{
data(sdata)
data(pathway)

fit.palmer <- palmer(X=sdata,path=pathway,K=2,L=3,B=100)

fit.palmer
predict(fit.palmer)
plot(fit.palmer)
}
