% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pct}
\alias{pct}
\title{Convert a vector of numbers to percentages.}
\usage{
pct(v, digits = 0)
}
\arguments{
\item{v}{A vector of numbers.}

\item{digits}{The number of digits after the decimal. Default is zero.}
}
\value{
A character vector.
}
\description{
Convert numbers to percentags. If all numbers are less than 1, numbers are
multiplied by 100. If any number is greater than 1, numbers are not multiplied.
}
\examples{
pct(.5)
pct(runif(10), digits=2)
pct(100*runif(10))

}
