% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{percent}
\alias{percent}
\title{Summarize a data frame and calculate percent for each group.}
\usage{
percent(.data, ..., .keep_n = F)
}
\arguments{
\item{.data}{A data frame or tibble.}

\item{...}{Grouping variables.}

\item{.keep_n}{Boolean to keep the counts in addition to the percentage. Default is FALSE.}
}
\value{
A data frame.
}
\description{
Summarize a data frame and calculate percent for each group.
}
\examples{
percent(mtcars, cyl)
percent(mtcars, cyl, gear)

}
