% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmer_method.R
\docType{methods}
\name{plot,palmer,missing-method}
\alias{plot,palmer,missing-method}
\alias{plot.palmer}
\title{Plotting function}
\usage{
\S4method{plot}{palmer,missing}(x, y, ...)
}
\arguments{
\item{x}{output of palmer function}
}
\description{
Heatmap with clustered data
}
