% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pr}
\alias{pr}
\title{Print as part of pipe. Print and return original value.}
\usage{
pr(x)
}
\arguments{
\item{x}{Any printable object.}
}
\value{
x
}
\description{
Print as part of pipe. Print and return original value.
}
