\name{predict,palmer-method}
\alias{predict,palmer-method}
\alias{predict.palmer}
\docType{methods}

\title{palmer result}
\usage{
\S4method{predict}{palmer}(object)
}
\arguments{
\item{object}{Output of palmer function}
}
\description{
Provide Gene and GO term cluster and its probability.
}
