% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{presidents}
\alias{presidents}
\title{U.S. Presidents}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 46 rows and 18 columns.
}
\source{
N/A
}
\usage{
presidents
}
\description{
U.S. Presidents
}
\keyword{datasets}
