% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportional_facets.R
\name{proportional_facet_heights}
\alias{proportional_facet_heights}
\title{Proportional facet heights}
\usage{
proportional_facet_heights(p)
}
\arguments{
\item{p}{A ggplot with wrapped facets in a single column using \code{facet_wrap(~ var, ncol=1, scales="free_y")}

#' @import ggplot2, grid, dplyr}
}
\value{
A grob
}
\description{
The default behavior for ggplot facets is that all facets have the same dimension. When plotting
facets with discrete axes and \code{scales=free_y}, the axis spacing will be inconsistent. This function
adjusts the relative heights of the facets so that the facet heights are proportional to the number
of items on the discrete scale for each facet.
}
