% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_theme.R
\name{rcols.pallette}
\alias{rcols.pallette}
\title{Reasonable Colors palette generator}
\usage{
rcols.pallette(color, shade = 3, pairing = "complementary")
}
\arguments{
\item{color}{Starting color for palette. One of "rose", "raspberry", "red", "orange", "cinnamon", "amber", "yellow", "lime", "chartreuse", "green", "emerald", "aquamarine", "teal", "cyan", "powder", "sky", "cerulean", "azure", "blue", "indigo", "violet", "purple", "magenta", "pink".}

\item{shade}{An integer from 1 (lightest) to 6 (darkest). Default is 3.}

\item{pairing}{One of "complementary", "split_complementary", "triadic", "tetradic", "analogous", or "set6".}
}
\value{
A named vector of html color codes.
}
\description{
Reasonable Colors palette generator
}
\examples{
rcols.pallette("green", shade=2, pairing="triadic")

}
