% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rcols}
\alias{rcols}
\title{The "Reasonable Colors" palette developed by Matthew Howell. https://www.reasonable.work/colors/}
\format{
An object of class \code{character} of length 150.
}
\usage{
rcols
}
\description{
The "Reasonable Colors" palette developed by Matthew Howell. https://www.reasonable.work/colors/
}
\keyword{datasets}
