% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{renumber_polygons}
\alias{renumber_polygons}
\title{Renumber polygons so neighboring polygons are more likely to have sequential numbers.}
\usage{
renumber_polygons(shp, adj = NULL)
}
\arguments{
\item{shp}{A polygon shape file to renumber.}

\item{adj}{Optional. An adjacency list, as produced by \code{palmer::adjacent_polys}.}
}
\value{
A vector of new district numbers
}
\description{
Renumber polygons so neighboring polygons are more likely to have sequential numbers.
}
