% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmer_method.R
\docType{methods}
\name{show,palmer-method}
\alias{show,palmer-method}
\alias{show.palmer}
\title{Summary of palmer results}
\usage{
\S4method{show}{palmer}(object)
}
\arguments{
\item{object}{output of palmer function}
}
\description{
Summary of palmer results
}
