% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{st_envelope}
\alias{st_envelope}
\title{Get envelope of a shape file}
\usage{
st_envelope(shp)
}
\arguments{
\item{shp}{A sf object}
}
\value{
A sf object.
}
\description{
Get envelope of a shape file
}
\examples{
st_envelope(ma[1,])

}
