% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{st_rook}
\alias{st_rook}
\title{Calculate rook adjacency}
\usage{
st_rook(a, b = a)
}
\arguments{
\item{a}{A sf object.}

\item{b}{A sf object. Optional. Default is object \code{a}.}
}
\value{
Sparse geometry binary predicate list
}
\description{
Calculate rook adjacency
}
\examples{
st_rook(ma)

}
