% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{st_weighted_centroid}
\alias{st_weighted_centroid}
\title{Find the weighted centroid of a set of points}
\usage{
st_weighted_centroid(x, weight)
}
\arguments{
\item{x}{A sf or sfc object. Must be "POINTS".}

\item{weight}{A vector of weights}
}
\value{
A sfc object with the weighted centroid of the points
}
\description{
Find the weighted centroid of a set of points
}
