% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_lookup}
\alias{state_lookup}
\title{Data on U.S. States}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 51 rows and 19 columns.
}
\usage{
state_lookup
}
\description{
Data on U.S. States
}
\keyword{datasets}
