% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_codes.R
\name{state_std}
\alias{state_std}
\title{Standardize states from state names, abbreviations, or fips codes.}
\usage{
state_std(s = NULL, type = "fips")
}
\arguments{
\item{s}{A state name, two-letter abbreviate, or FIPS code, or a vector}

\item{type}{Output type. "fips", "abb", or "name". Default is "fips".}
}
\description{
Standardize states from state names, abbreviations, or fips codes.
}
\examples{
state_std(c("MA", "NY", "California", "025", "33", 34), type="fips")
state_std(c("MA", "NY", "California", "025", "33", 34), type="abb")
state_std(c("MA", "NY", "California", "025", "33", 34), type="name")

}
