% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_theme.R
\name{theme_pres}
\alias{theme_pres}
\title{ggplot2 theme for presentations}
\usage{
theme_pres(base_size = 32, base_family = "Montserrat")
}
\arguments{
\item{base_size}{Base font size. Default is 32.}

\item{base_family}{Base font family. Default is Montserrat (must be installed)}
}
\value{
ggplot2 theme
}
\description{
ggplot2 theme for presentations
}
