% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{widest_rectangle}
\alias{widest_rectangle}
\title{Find widest rectangle in state.}
\usage{
widest_rectangle(shp, n = c(20, 20))
}
\arguments{
\item{shp}{Polygon shape file}

\item{n}{Two-item list of the number of rectangles to draw in each direction. Higher values are more precise but slower. Default is \code{c(20,20)}.}
}
\value{
Centroid of wide rectangle.
}
\description{
Used to label placement. Finds a rectangle at least 90\% as wide as possible, with preference for rectangles closer to the centroid.
}
