% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{palmer-class}
\alias{palmer-class}
\title{An S4 class to represent hubviz model fitting results.}
\description{
An S4 class to represent hubviz model fitting results.
}
\section{Slots}{

\describe{
\item{\code{data}}{data}

\item{\code{init}}{model initialization}

\item{\code{result}}{clustering analysis results}
}}

