
\name{plot,palmer,missing-method}
\alias{plot,palmer,missing-method}
\alias{plot.palmer}
\docType{methods}

\title{Plotting function}
\usage{
\S4method{plot}{palmer,missing}(x, y, ...)
}
\arguments{
\item{x}{Output of palmer function}
}
\description{
Provide visualization of constrained biclustersing analysis result.
}
