% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmer_method.R
\docType{methods}
\name{predict,palmer-method}
\alias{predict,palmer-method}
\alias{predict.palmer}
\alias{#'}
\title{palmer result}
\usage{
\S4method{predict}{palmer}(object)
}
\arguments{
\item{object}{output of palmer function}
}
\description{
Return palmer result
}
