% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{qmap}
\alias{qmap}
\title{Make a simple ggplot map}
\usage{
qmap(.data, size = 0.25, ...)
}
\arguments{
\item{.data}{A sf object}

\item{size}{Optional point size}

\item{...}{Additional arguments passed to \code{geom_sf}}
}
\value{
A ggplot2 object
}
\description{
Make a simple ggplot map
}
